Extract chrome-headless-shell path utilities to break circular dependency
Move path/version utility functions from chrome-headless-shell.ts into
chrome-headless-shell-paths.ts so puppeteer.ts can import them without
creating a cycle (chrome-headless-shell → chromium → puppeteer →
chrome-headless-shell). The original module re-exports everything for
backward compatibility.
