+++ src/tools/impl/chrome-headless-shell.ts
@@ -10,7 +10,6 @@
 
 import { join } from "../../deno_ral/path.ts";
 import { existsSync, safeMoveSync, safeRemoveSync } from "../../deno_ral/fs.ts";
-import { quartoDataDir } from "../../core/appdirs.ts";
 import {
   InstallableTool,
   InstallContext,
@@ -22,65 +21,27 @@ import {
   downloadAndExtractChrome,
   fetchLatestCftRelease,
   fetchPlaywrightBrowsersJson,
-  findChromeExecutable,
   isPlaywrightCdnPlatform,
   playwrightCdnDownloadUrl,
 } from "./chrome-for-testing.ts";
+import {
+  chromeHeadlessShellBinaryName,
+  chromeHeadlessShellInstallDir,
+  isInstalled,
+  noteInstalledVersion,
+  readInstalledVersion,
+} from "./chrome-headless-shell-paths.ts";
+import { chromiumInstallable } from "./chromium.ts";
 
-const kVersionFileName = "version";
-
-// -- Version helpers --
-
-/** Return the chrome-headless-shell install directory under quartoDataDir. */
-export function chromeHeadlessShellInstallDir(): string {
-  return quartoDataDir("chrome-headless-shell");
-}
-
-/**
- * The executable name for chrome-headless-shell on the current platform.
- * CfT builds use "chrome-headless-shell", Playwright arm64 builds use "headless_shell".
- * Returns the CfT name if platform detection fails (unsupported platform).
- */
-export function chromeHeadlessShellBinaryName(): string {
-  try {
-    return isPlaywrightCdnPlatform() ? "headless_shell" : "chrome-headless-shell";
-  } catch {
-    return "chrome-headless-shell";
-  }
-}
-
-/**
- * Find the chrome-headless-shell executable in the install directory.
- * Returns the absolute path if installed, undefined otherwise.
- */
-export function chromeHeadlessShellExecutablePath(): string | undefined {
-  const dir = chromeHeadlessShellInstallDir();
-  if (!existsSync(dir)) {
-    return undefined;
-  }
-  return findChromeExecutable(dir, chromeHeadlessShellBinaryName());
-}
-
-/** Record the installed version as a plain text file. */
-export function noteInstalledVersion(dir: string, version: string): void {
-  Deno.writeTextFileSync(join(dir, kVersionFileName), version);
-}
-
-/** Read the installed version. Returns undefined if not present. */
-export function readInstalledVersion(dir: string): string | undefined {
-  const path = join(dir, kVersionFileName);
-  if (!existsSync(path)) {
-    return undefined;
-  }
-  const text = Deno.readTextFileSync(path).trim();
-  return text || undefined;
-}
-
-/** Check if chrome-headless-shell is installed in the given directory. */
-export function isInstalled(dir: string): boolean {
-  return existsSync(join(dir, kVersionFileName)) &&
-    findChromeExecutable(dir, chromeHeadlessShellBinaryName()) !== undefined;
-}
+// Re-export path utilities for external consumers
+export {
+  chromeHeadlessShellBinaryName,
+  chromeHeadlessShellExecutablePath,
+  chromeHeadlessShellInstallDir,
+  isInstalled,
+  noteInstalledVersion,
+  readInstalledVersion,
+} from "./chrome-headless-shell-paths.ts";
 
 // -- InstallableTool methods --
 
@@ -180,7 +141,19 @@ async function install(pkg: PackageInfo, _ctx: InstallContext): Promise<void> {
   noteInstalledVersion(installDir, pkg.version);
 }
 
-async function afterInstall(_ctx: InstallContext): Promise<boolean> {
+async function afterInstall(ctx: InstallContext): Promise<boolean> {
+  // Clean up legacy chromium installed by 'quarto install chromium'
+  try {
+    if (await chromiumInstallable.installed()) {
+      ctx.info("Removing legacy Chromium installation...");
+      await chromiumInstallable.uninstall(ctx);
+    }
+  } catch {
+    ctx.info(
+      "Note: Could not remove legacy Chromium. " +
+        "You can remove it manually with 'quarto uninstall chromium'.",
+    );
+  }
   return false;
 }
